pyproject.toml
[tool.poetry]
name = "GPU Training Course"
version = "0.1.0"
description = "A training course to explain the use of GPUs"
authors = ["Liam Berrisford <l.berrisford3@exeter.ac.uk>"]
readme = "README.md"
packages = [{include = "content"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
xarray = "^2023.5.0"       # For handling NetCDF data
netCDF4 = "^1.6.1"         # For reading NetCDF files
matplotlib = "^3.7.1"      # For plotting
plotly = "^5.15.0"         # For interactive plotting and HTML export
cupy-cuda12x = "*"         # For CUDA support with CuPy
copernicusmarine = "*"     # For Copernicus Marine Service data access
tqdm = "*"
jupyter-book = "^1.0.0"

[tool.poetry.scripts]
visualise_slice_static = "content.support_scripts:visualisation_static"
visualise_slice = "content.support_scripts:visualisation_slice"
visualise_cube = "content.support_scripts:visualisation_cube"
summary =  "content.support_scripts:summary"
cuda_check = "content.support_scripts:cuda_check"
download_data = "content.support_scripts:download_ocean_data"
diffusion_cupy = "content.temperature_diffusion:run_diffusion_cupy"
diffusion_pure_python = "content.temperature_diffusion:run_diffusion_purepython"
diffusion_numpy = "content.temperature_diffusion:run_diffusion_numpy"
game_of_life_cpu = "content.game_of_life:run_life_numpy"
game_of_life_gpu = "content.game_of_life:run_life_cupy"
game_of_life_naive = "content.game_of_life:run_life_naive"
game_of_life_cpu_profiled = "content.game_of_life_profiled:run_life_numpy"
game_of_life_gpu_profiled = "content.game_of_life_profiled:run_life_cupy"
game_of_life_naive_profiled = "content.game_of_life_profiled:run_life_naive"
game_of_life_experiment_profiled = "content.game_of_life_experiment_profiled:run_experiment"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
